+++ KeeForge/Models/EntryEdit.swift
@@ -3,17 +3,20 @@ import Foundation
 struct EntryDraftPayload: Codable, Sendable, Equatable {
     struct TOTPConfiguration: Codable, Sendable, Equatable {
         var secret: String
+        var decodedSecret: Data?
         var period: Int
         var digits: Int
         var algorithm: TOTPAlgorithm
 
         init(
             secret: String,
+            decodedSecret: Data? = nil,
             period: Int = 30,
             digits: Int = 6,
             algorithm: TOTPAlgorithm = .sha1
         ) {
             self.secret = secret
+            self.decodedSecret = decodedSecret
             self.period = period
             self.digits = digits
             self.algorithm = algorithm